+++ src/tools.js
@@ -13,22 +13,22 @@ Tools.install = (Vue, options) => {
         },
 
         // 是否安装微信
-        isInstallWXApp () {
-            return new Promise((resolve, reject) => {
-                tools.isInstallWXApp(({ status, errorMsg, data }) => {
-                    status === 0 ? resolve(data) : reject({ status, errorMsg, data })
-                })
-            })
-        },
+        // isInstallWXApp () {
+        //     return new Promise((resolve, reject) => {
+        //         tools.isInstallWXApp(({ status, errorMsg, data }) => {
+        //             status === 0 ? resolve(data) : reject({ status, errorMsg, data })
+        //         })
+        //     })
+        // },
 
-        // 获取 cid
-        getCid () {
-            return new Promise((resolve, reject) => {
-                tools.getCid(({ status, errorMsg, data }) => {
-                   status === 0 ? resolve(data) : reject({ status, errorMsg, data })
-                })
-            })
-        },
+        // // 获取 cid
+        // getCid () {
+        //     return new Promise((resolve, reject) => {
+        //         tools.getCid(({ status, errorMsg, data }) => {
+        //            status === 0 ? resolve(data) : reject({ status, errorMsg, data })
+        //         })
+        //     })
+        // },
 
         // 复制内容到剪切板
         copyString (string) {